+++ lib/classes/API.ts
@@ -1,7 +1,9 @@
 import fetch from "node-fetch";
+import { Endpoints } from "../structures/enums/Endpoints";
 import { MeResponse } from "../structures/MeResponse";
 import { SearchParameters } from "../structures/SearchParameters";
 import { SearchResponse } from "../structures/SearchResponse";
+import { Utils } from "./Utils";
 
 export class API {
 
@@ -11,30 +13,57 @@ export class API {
     };
 
     /**
+     * The utils class.
+     */
+    public utils: Utils;
+
+    /**
      * Constructs an instance of the API.
      */
     constructor() {
         this.uris = {
-            search: `https://api.trace.moe/search`,
-            me: `https://api.trace.moe/me`
+            search: Endpoints.SEARCH,
+            me: Endpoints.ME
         };
+        this.utils = new Utils();
     }
    
     /**
      * Searches the website for the similar anime.
      * @param {string} imageURL The URL for the image.
-     * @param {SearchParameters} extras Extra parameters for the search.
+     * @param {SearchParameters} options Optional parameters for the search.
      * @returns {SearchResponse}
      */
-    async fetchAnime(imageURL: string, { cutBorders = false, anilistId = false }: SearchParameters = {}): Promise<SearchResponse> {
-        let url = `${this.uris.search}?url=${encodeURIComponent(imageURL)}&anilistInfo`;
-        if (cutBorders)
-            url += `&cutBorders`;
-        
-        if (anilistId)
-            url += `&anilistID=1`;
+    async fetchAnime(imageURL: string, options?: SearchParameters): Promise<SearchResponse> {
+        if (!(imageURL) || typeof imageURL !== "string")
+            throw new TypeError(`'imageURL' should be type string. Got type ${typeof imageURL} instead.`);
+
+        const url = this.utils.handleOptionalParameters(this.uris.search, imageURL, options);
 
         return await fetch(url)
+            .then((res) => res.json())
+            .catch((err) => { throw new Error(err)} );
+    }
+
+    /**
+     * Searches the website for the similar anime but this time using an image file.
+     * @param {Buffer} buffer The image buffer. Limited to 25 MB.
+     * @param {SearchParameters} options Optional parameters for the search.
+     * @returns {SearchResponse}
+     */
+    async fetchAnimeFromBuffer(buffer: Buffer, options?: SearchParameters): Promise<SearchResponse> {
+        if (!(buffer) || !(buffer instanceof Buffer))
+            throw new TypeError(`'buffer' should be type string. Got type ${typeof buffer} instead.`);
+
+        const url = this.utils.handleOptionalParameters(this.uris.search, undefined, options);
+
+        return await fetch(url, {
+            method: "POST",
+            body: buffer,
+            headers: {
+                "Content-Type": "application/x-www-form-urlencoded"
+            }
+        })
             .then((res) => res.json())
             .catch((err) => { throw new Error(err)} );
     }