+++ lib/structures/enums/Endpoints.ts
@@ -1,0 +1,4 @@
+export enum Endpoints {
+    SEARCH = "https://api.trace.moe/search",
+    ME = "https://api.trace.moe/me"
+}