+++ tests/api.spec.ts
@@ -1,16 +1,39 @@
 import { expect } from 'chai';
+import { readFileSync } from "fs";
 import { TraceMoe } from '../lib/index';
 const api = new TraceMoe();
 
 describe("API#fetchAnime", () => {
     let result: { anilist: { id: any; }; filename: any; episode: any; similarity: any; video: any; image: any; };
     it("should return an anime object", async () => {
-        const res = await api.fetchAnime("https://cdn.xndr.tech/u/KdU9oT3.jpeg");
+        const res = await api.fetchAnime("https://cdn.xndr.tech/u/KdU9oT3.jpeg", { anilistInfo: true });
+        console.log(res.result[0]);
+        
         result = res.result[0];
     })
         .timeout(7500);
     it("result properties should exist", () => {
         expect(result.anilist.id).to.be.a("number");
+        expect(result.filename).to.be.a("string");
+        expect(result.episode).to.be.a("number");
+        expect(result.similarity).to.be.a("number");
+        expect(result.video).to.be.a("string");
+        expect(result.image).to.be.a("string");
+    });
+});
+
+describe("API#fetchAnimeFromBuffer", () => {
+    let result: { anilist: { id: any; }; filename: any; episode: any; similarity: any; video: any; image: any; };
+    it("buffer example should be an valid buffer", () => {
+        Buffer.isBuffer(readFileSync(__dirname + "/assets/bufferExample.jpg"));
+    })
+    it("should return an anime object", async () => {
+        const res = await api.fetchAnimeFromBuffer(readFileSync(__dirname + "/assets/bufferExample.jpg"));
+        result = res.result[0];
+    })
+        .timeout(7500);
+    it("result properties should exist", () => {
+        expect(result.anilist).to.be.a("number");
         expect(result.filename).to.be.a("string");
         expect(result.episode).to.be.a("number");
         expect(result.similarity).to.be.a("number");